aegis_sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-secure"
version = "2.0.0"
description = "Governance guardrails for AI agents — policy decorators, real-time audit logging, kill-switch, and human-in-the-loop approval via MongoDB."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Aegis Team", email = "aegis.sentinel.dev@gmail.com"},
]
keywords = [
    "ai", "agents", "guardrails", "governance", "safety",
    "langchain", "llm", "policy", "audit", "decorator",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
langchain = ["langchain>=0.2", "langgraph>=0.2"]
dev = ["build", "twine", "pytest"]

[project.urls]
Homepage = "https://github.com/Baalavignesh/Aegis"
Repository = "https://github.com/Baalavignesh/Aegis"
Documentation = "https://github.com/Baalavignesh/Aegis#readme"
"Bug Tracker" = "https://github.com/Baalavignesh/Aegis/issues"

[project.scripts]
sentinel = "sentinel.cli:main"

[tool.setuptools]
license-files = []

[tool.setuptools.packages.find]
include = ["sentinel*"]
